requirements.txt
# LLM API clients
google-generativeai>=0.3.0
openai>=1.0.0

# YAML processing
pyyaml>=6.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
